Move ScriptCheckTest from server_test to runtime_test
