+++ lib/tests/streamlit/web/server/server_test.py
@@ -17,8 +17,6 @@
 import asyncio
 import errno
 import os
-import shutil
-import tempfile
 import unittest
 from unittest import mock
 from unittest.mock import patch
@@ -36,7 +34,6 @@ from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.runtime.forward_msg_cache import populate_hash_if_needed
 from streamlit.runtime.runtime import RuntimeState
 from streamlit.runtime.uploaded_file_manager import UploadedFileRec
-from streamlit.watcher import event_based_path_watcher
 from streamlit.web.server.server import (
     MAX_PORT_SEARCH_RETRIES,
     RetriesExceeded,
@@ -529,65 +526,6 @@ class UnixSocketTest(unittest.TestCase):
                 "/home/superfakehomedir/fancy-test/testasd"
             )
             mock_server.add_socket.assert_called_with(some_socket)
-
-
-@patch("streamlit.source_util._cached_pages", new=None)
-class ScriptCheckTest(tornado.testing.AsyncTestCase):
-    def setUp(self) -> None:
-        super().setUp()
-
-        self._home = tempfile.mkdtemp()
-        self._old_home = os.environ["HOME"]
-        os.environ["HOME"] = self._home
-
-        self._fd, self._path = tempfile.mkstemp()
-        self._server = Server(self._path, "test command line")
-        self._server._runtime._eventloop = self.asyncio_loop
-
-    def tearDown(self) -> None:
-        if event_based_path_watcher._MultiPathWatcher._singleton is not None:
-            event_based_path_watcher._MultiPathWatcher.get_singleton().close()
-            event_based_path_watcher._MultiPathWatcher._singleton = None
-
-        os.environ["HOME"] = self._old_home
-        os.remove(self._path)
-        shutil.rmtree(self._home)
-
-        super().tearDown()
-
-    @pytest.mark.slow
-    @tornado.testing.gen_test(timeout=30)
-    async def test_invalid_script(self):
-        await self._check_script_loading(
-            "import streamlit as st\n\nst.deprecatedWrite('test')",
-            False,
-            "error",
-        )
-
-    @pytest.mark.slow
-    @tornado.testing.gen_test(timeout=30)
-    async def test_valid_script(self):
-        await self._check_script_loading(
-            "import streamlit as st\n\nst.write('test')", True, "ok"
-        )
-
-    @pytest.mark.slow
-    @tornado.testing.gen_test(timeout=30)
-    async def test_timeout_script(self):
-        with patch("streamlit.runtime.runtime.SCRIPT_RUN_CHECK_TIMEOUT", new=0.1):
-            await self._check_script_loading(
-                "import time\n\ntime.sleep(5)", False, "timeout"
-            )
-
-    async def _check_script_loading(self, script, expected_loads, expected_msg):
-        with os.fdopen(self._fd, "w") as tmp:
-            tmp.write(script)
-
-        ok, msg = await self._server._runtime.does_script_run_without_error()
-        event_based_path_watcher._MultiPathWatcher.get_singleton().close()
-        event_based_path_watcher._MultiPathWatcher._singleton = None
-        self.assertEqual(expected_loads, ok)
-        self.assertEqual(expected_msg, msg)
 
 
 class ScriptCheckEndpointExistsTest(tornado.testing.AsyncHTTPTestCase):
